Added registration

+++ src/main/java/com/crnjakovic/controller/AppController.java
@@ -1,6 +1,7 @@
 package com.crnjakovic.controller;
 
 import com.crnjakovic.model.Game;
+import com.crnjakovic.model.Player;
 import com.crnjakovic.service.GameService;
 import com.crnjakovic.service.PlayerService;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -8,8 +9,13 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
+import org.springframework.validation.BindingResult;
+import org.springframework.validation.Errors;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.context.request.WebRequest;
 import org.springframework.web.servlet.ModelAndView;
+
+import javax.validation.Valid;
 
 /**
  * Created by lukacrnjakovic on 3/24/18.
@@ -58,7 +64,9 @@ public class AppController {
         Authentication auth = SecurityContextHolder.getContext().getAuthentication();
         String name = auth.getName();
         if(attr.equals("left")){
-            gameService.deleteGame(name);
+            if(gameService.checkIfUserPlaying(name)){
+                gameService.deleteGame(name);
+            }
         }
         return "redirect:home";
     }
@@ -87,5 +95,36 @@ public class AppController {
         }
     }
 
+    @RequestMapping(value = "register", method = RequestMethod.GET)
+    public String showRegistrationForm(WebRequest request, Model model) {
+        UserDto user = new UserDto();
+        model.addAttribute("user", user);
+        return "register";
+    }
+
+    @PostMapping(value = "registering")
+    public ModelAndView registerUserAccount(
+            @ModelAttribute("user") @Valid UserDto accountDto,
+            BindingResult result,
+            WebRequest request,
+            Errors errors) {
+        if(playerService.findPlayerByUsername(accountDto.getUserName())!=null){
+            return new ModelAndView("register", "user", accountDto);
+        }
+        if (!result.hasErrors()) {
+            createUserAccount(accountDto, result);
+        }
+        if (result.hasErrors()) {
+            return new ModelAndView("register", "user", accountDto);
+        }
+        else {
+            return new ModelAndView("redirect:/app/secure/home");
+        }
+    }
+    private Player createUserAccount(UserDto accountDto, BindingResult result) {
+        Player registered = null;
+        registered = playerService.registerNewUserAcount(accountDto);
+        return registered;
+    }
 }
